Remove features module entirely
