+++ gneiss-mqtt/src/client/config.rs
@@ -16,6 +16,8 @@ use std::fmt::{Debug, Formatter};
 use std::fs::File;
 use std::io::Read;
 use std::net::{SocketAddr, ToSocketAddrs};
+#[cfg(any(feature = "tokio-native-tls", feature = "threaded-native-tls", feature = "tokio-rustls", feature = "threaded-rustls"))]
+use std::sync::Arc;
 use std::time::Duration;
 
 #[cfg(any(feature="tokio-websockets", feature="threaded-websockets"))]
@@ -24,6 +26,8 @@ use http::{Uri, Version};
 use std::str::FromStr;
 #[cfg(feature="tokio-websockets")]
 use std::{future::Future, pin::Pin,};
+#[cfg(any(feature = "tokio-rustls", feature = "threaded-rustls"))]
+use rustls_pki_types::PrivateKeyDer;
 #[cfg(any(feature="tokio-websockets", feature="threaded-websockets"))]
 use tungstenite::{client::*, handshake::client::generate_key};
 
@@ -278,6 +282,138 @@ impl TlsOptionsBuilder {
         self
     }
 
+    #[cfg(any(feature = "tokio-native-tls", feature = "threaded-native-tls"))]
+    /// Builds client TLS options using the `native-tls` crate
+    ///
+    /// If using MTLS, native-tls only supports pkcs8 format private keys.  If your private key is in a different
+    /// format, you must first convert it to pkcs8 and instead use that.
+    pub fn build_native_tls(&self) -> Result<TlsOptions, MqttError> {
+        let mut builder = native_tls::TlsConnector::builder();
+
+        if let Some(root_ca_bytes) = &self.root_ca_bytes {
+            builder.disable_built_in_roots(true);
+
+            let root = native_tls::Certificate::from_pem(root_ca_bytes.as_slice())?;
+            builder.add_root_certificate(root);
+        }
+
+        if self.mode == TlsMode::Mtls {
+            let identity = native_tls::Identity::from_pkcs8(self.certificate_bytes.as_ref().unwrap(), self.private_key_bytes.as_ref().unwrap())?;
+            builder.identity(identity);
+        }
+
+        if let Some(alpn) = &self.alpn {
+            let protocols = vec!(alpn.as_str());
+            builder.request_alpns(protocols.as_slice());
+        }
+
+        builder.use_sni(self.verify_peer);
+
+        Ok(TlsOptions {
+            options: TlsData::NativeTls(Arc::new(builder))
+        })
+    }
+
+    #[cfg(any(feature = "tokio-native-tls", feature = "threaded-native-tls"))]
+    /// Builds client TLS options directly from a native-tls TlsConnectorBuilder instance
+    ///
+    /// This factory is useful when you need TLS properties that TlsOptionsBuilder does not support.
+    pub fn build_native_tls_from_tls_connector_builder(builder: native_tls::TlsConnectorBuilder) -> TlsOptions {
+        TlsOptions {
+            options: TlsData::NativeTls(Arc::new(builder))
+        }
+    }
+}
+
+
+impl TlsOptionsBuilder {
+
+    #[cfg(any(feature = "tokio-rustls", feature = "threaded-rustls"))]
+    /// Builds client TLS options using the `rustls` crate
+    pub fn build_rustls(&self) -> Result<TlsOptions, MqttError> {
+        let root_cert_store = build_root_ca_store(self.root_ca_bytes.as_deref())?;
+
+        let mut config =
+            match self.mode {
+                TlsMode::Standard => {
+                    rustls::ClientConfig::builder()
+                        .with_root_certificates(root_cert_store)
+                        .with_no_client_auth()
+                }
+                TlsMode::Mtls => {
+                    let certs = build_certs(self.certificate_bytes.as_deref().unwrap());
+                    let private_key = build_private_key(self.private_key_bytes.as_deref().unwrap())?;
+                    rustls::ClientConfig::builder()
+                        .with_root_certificates(root_cert_store)
+                        .with_client_auth_cert(certs, private_key)?
+                }
+            };
+
+        config.alpn_protocols = Vec::new();
+        if let Some(alpn) = &self.alpn {
+            config.alpn_protocols.push(alpn.as_bytes().to_vec());
+        }
+
+        config.enable_sni = self.verify_peer;
+
+        Ok(TlsOptions {
+            options: TlsData::Rustls(Arc::new(config))
+        })
+    }
+
+    #[cfg(any(feature = "tokio-rustls", feature = "threaded-rustls"))]
+    /// Builds client TLS options directly from a Rustls ClientConfig instance
+    ///
+    /// This factory is useful when you need TLS properties that TlsOptionsBuilder does not support.
+    pub fn build_rustls_from_client_config(config: rustls::ClientConfig) -> TlsOptions {
+        TlsOptions {
+            options: TlsData::Rustls(Arc::new(config))
+        }
+    }
+}
+
+#[cfg(any(feature = "tokio-rustls", feature = "threaded-rustls"))]
+fn build_root_ca_store(root_ca_bytes: Option<&[u8]>) -> Result<rustls::RootCertStore, std::io::Error> {
+    let mut root_cert_store = rustls::RootCertStore::empty();
+    if let Some(root_ca_bytes) = root_ca_bytes {
+        let mut pem = std::io::BufReader::new(root_ca_bytes);
+        for cert in rustls_pemfile::certs(&mut pem) {
+            root_cert_store.add(cert?).unwrap();
+        }
+    } else {
+        let system_certs = rustls_native_certs::load_native_certs();
+        for cert in system_certs.unwrap() {
+            root_cert_store.add(cert).unwrap();
+        }
+    }
+
+    Ok(root_cert_store)
+}
+
+#[cfg(any(feature = "tokio-rustls", feature = "threaded-rustls"))]
+fn build_private_key(key_bytes: &[u8]) -> Result<PrivateKeyDer<'static>, rustls::Error> {
+    let mut reader = std::io::BufReader::new(key_bytes);
+
+    loop {
+        let read_pem_result = rustls_pemfile::read_one(&mut reader);
+        match read_pem_result {
+            Ok(Some(rustls_pemfile::Item::Pkcs1Key(key))) => return Ok(key.into()),
+            Ok(Some(rustls_pemfile::Item::Pkcs8Key(key))) => return Ok(key.into()),
+            Ok(Some(rustls_pemfile::Item::Sec1Key(key))) => return Ok(key.into()),
+            Ok(None) => { return Err(rustls::Error::General("no valid private keys found".to_string())); }
+            Ok(_) => {}
+            Err(_) => { return Err(rustls::Error::General("failed to parse private key pem file".to_string())); }
+        }
+    }
+}
+
+#[cfg(any(feature = "tokio-rustls", feature = "threaded-rustls"))]
+fn build_certs(certificate_bytes: &[u8]) -> Vec<rustls_pki_types::CertificateDer<'static>> {
+    let mut reader = std::io::BufReader::new(certificate_bytes);
+
+    rustls_pemfile::certs(&mut reader)
+        .flatten()
+        .collect()
 }
 
 impl Default for TlsOptionsBuilder {
